+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Core: Set initial PositionSet capacity to improve performance
 * Core: Fixed broken BoardMetrics calculations
 * Core: Empty board should always be scored as a zero
+* Core: Updated DefaultEngineConfig.xml with latest mergetop results
 * Test: Added TreeStrap tests
 * Trainer: AutoTrain now supports using MzingaAutoTrainConfig.xml